refactor: remove `kernel` module
To use this module with the initrd.
