+++ bootx64/src/fs/size.rs
