+++ bootx64/src/main.rs
@@ -22,7 +22,6 @@ mod mem;
 
 use common::{kernelboot, mem::reserved};
 use core::{convert::TryInto, ptr, ptr::NonNull, slice};
-use fs::kernel;
 use mem::{paging, stack};
 use uefi::{
     prelude::{Boot, Handle, SystemTable},
@@ -37,9 +36,9 @@ pub fn efi_main(image: Handle, system_table: SystemTable<Boot>) -> ! {
 
     let vram_info = gop::init(system_table.boot_services());
 
-    let (phys_kernel_addr, bytes_kernel) = kernel::deploy(system_table.boot_services());
+    let (phys_kernel_addr, bytes_kernel) = fs::deploy(system_table.boot_services());
     let (entry_addr, actual_mem_size) =
-        kernel::fetch_entry_address_and_memory_size(phys_kernel_addr, bytes_kernel);
+        fs::fetch_entry_address_and_memory_size(phys_kernel_addr, bytes_kernel);
 
     let stack_addr = stack::allocate(system_table.boot_services());
     let reserved_regions = reserved::Map::new(
